Tabele kontrole

+++ database/migrations/2023_11_11_125044_create_porezdoprinosis_table.php
@@ -14,32 +14,32 @@ return new class extends Migration
         Schema::create('porezdoprinosis', function (Blueprint $table) {
             $table->id();
             $table->string('M_G_mesec_godina')->nullable();
-            $table->decimal('IZN1_iznos_poreskog_oslobodjenja', 10, 4)->nullable();
+            $table->float('IZN1_iznos_poreskog_oslobodjenja', 10, 4)->nullable();
             $table->string('OPPOR_opis_poreza', 35)->nullable();
-            $table->decimal('P1_porez_na_licna_primanja', 10, 4)->nullable();
+            $table->float('P1_porez_na_licna_primanja', 10, 4)->nullable();
 
             // Radnik
 
             $table->string('OPIS1_opis_zdravstvenog_osiguranja_na_teret_radnika', 35)->nullable();
-            $table->decimal('ZDRO_zdravstveno_osiguranje_na_teret_radnika', 10, 4)->nullable();
+            $table->float('ZDRO_zdravstveno_osiguranje_na_teret_radnika', 10, 4)->nullable();
 
 
             $table->string('OPIS2_opis_pio_na_teret_radnika', 35)->nullable();
-            $table->decimal('PIO_pio_na_teret_radnika', 10, 4)->nullable();
+            $table->float('PIO_pio_na_teret_radnika', 10, 4)->nullable();
 
             $table->string('OPIS3_opis_osiguranja_od_nezaposlenosti_na_teret_radnika', 35)->nullable();
-            $table->decimal('ONEZ_osiguranje_od_nezaposlenosti_na_teret_radnika', 10, 4)->nullable();
+            $table->float('ONEZ_osiguranje_od_nezaposlenosti_na_teret_radnika', 10, 4)->nullable();
 
-            $table->decimal('UKDOPR_ukupni_doprinosi_na_teret_radnika', 10, 4)->nullable();
+            $table->float('UKDOPR_ukupni_doprinosi_na_teret_radnika', 10, 4)->nullable();
 
             // poslodavac
             $table->string('OPIS4_opis_zdravstvenog_osiguranja_na_teret_poslodavca', 35)->nullable();
-            $table->decimal('DOPRA_zdravstveno_osiguranje_na_teret_poslodavca', 10, 4)->nullable();
+            $table->float('DOPRA_zdravstveno_osiguranje_na_teret_poslodavca', 10, 4)->nullable();
 
             $table->string('OPIS5_opis_pio_na_teret_poslodavca', 35)->nullable();
-            $table->decimal('DOPRB_pio_na_teret_poslodavca', 11, 4)->nullable();
+            $table->float('DOPRB_pio_na_teret_poslodavca', 11, 4)->nullable();
 
-            $table->decimal('UKDOPP_ukupni_doprinosi_na_teret_poslodavca', 10, 4)->nullable();
+            $table->float('UKDOPP_ukupni_doprinosi_na_teret_poslodavca', 10, 4)->nullable();
             $table->timestamps();
         });
     }
